Move models.resolve to api.resolve_model
